refactor: DAO를 인터페이스로 추상화한다
